For nodes that OOM on v1.16, skip cleaning all slots at startup
When generating the index at startup, we track all the slots that are
dirty and add them to the list that `clean` uses.  When `clean` runs
this next time, there are lots of slots to clean, which causes memory to
spike while many accounts are loaded into memory.  This is a performance
optimization, as it ensures old accounts and slots are removed, which
improves accounts index and account storage utilization.

If a node with low memory on v1.16 is OOM-ing, we can skip adding these
dirty slots that were found during index generation to the list that
`clean` uses.  Everything will still be correct, and everything will
eventually be cleaned (worst case after an epoch's-worth of slots).